This is the first screen of our program
The buttons don't do anything yet, except display the message, "hello" at the bottom of the screen. First of all, I'm not sure if you'll be able to access the image when running your computer. Second of all, once you get the image on your computer, you might have to download the pil function. I have access to a tutorial if you need it, although it might not work because you still don't have python on your actual computer.