Basically completed screen 2
So I basically completed screen 2 of the program. I still need to add in the part where the user picks if it is one or two guests they are adding in. Plus the program should eventually say "Member has x amount of passes left for the month". I haven't figured out how to do that yet and will start working on the next screen because I think that will help me figure that out.